==> 06-interactive.Rmd <==
# Interactive component

<script src="https://d3js.org/d3.v7.js"></script>

<h2> Trend of food delivery revenue by channels </h2>

<button onclick="update(Platform)">Platform</button>
<button onclick="update(Restaurant)">Restaurant</button>

<div id="my_dataviz"></div>

<script src="scripts/revenue_trend.js"></script>

The interactive bar chart above depicts the revenue generated by various food delivery channels. It depicts the upward trend in revenue for both the platform and the restaurant. Because we use the same scale, you can look at the revenue trend over the last year and get a sense of the entire volume.
